ai-service/requirements-full.txt
# BookEqualizer AI Service - Complete Dependencies
# Production-ready with all Phase 2 features

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# EPUB Processing
beautifulsoup4==4.12.2
ebooklib==0.18
lxml==4.9.3

# Text Processing & NLP
nltk==3.8.1
regex==2023.10.3
spacy==3.7.2
textstat==0.7.3

# AI & ML
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2

# Database & Vector Storage
asyncpg==0.29.0
pgvector==0.2.4
sqlalchemy==2.0.23

# Caching & Performance
redis==5.0.1
cachetools==5.3.2

# Data Processing
pandas==2.1.4
pydantic==2.5.0
typing-extensions==4.8.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0